[FIX] not-so-special handling of zero-pupulated headers
At the end of the tar archive there are two 512-byte blocks filled with
binary zeros as an end-of-file marker.
We do not implement any special state machines to check adherence to this
specification. For simplicity of implementation and our own non-strong
requirements we ignore these blocks (which are usually headers).
This means that we could continue extracting files even after encountering
the two EOF marker blocks. (Ex. concatenation of two tar with the `cat`
command).